# docs/swagger.yaml
definitions:
...
  model.Post:
    properties:
      bid:
        type: string
      collectNum:
        type: integer
      commentNum:
        type: integer
      createdAt:
        type: string
      introduce:
        type: string
      likeNum:
        type: integer
      showImg:
        type: string
      studentID:
        type: string
      title:
        type: string
    type: object
  model.PostDraft:
    properties:
      bid:
        type: string
      createdAt:
        type: string
      introduce:
        type: string
      showImg:
        type: string
      studentID:
        type: string
      title:
        type: string
    type: object
...
  req.CreatePostReq:
    properties:
      introduce:
        type: string
      showImg:
        items:
          type: string
        type: array
      studentid:
        type: string
      title:
        type: string
    type: object
...
  req.FindPostReq:
    properties:
      name:
        type: string
    type: object
  req.InteractionReq:
    properties:
      studentid:
        type: string
      subject:
        type: string
      targetid:
        type: string
    type: object
...
  resp.ListPostsResp:
    properties:
      bid:
        type: string
      collectNum:
        type: integer
      commentNum:
        type: integer
      introduce:
        type: string
      isCollect:
        type: string
      isLike:
        type: string
      likeNum:
        type: integer
      showImg:
        items:
          type: string
        type: array
      title:
        type: string
      userInfo:
        properties:
          avatar:
            type: string
          school:
            type: string
          studentid:
            type: string
          username:
            type: string
        type: object
    type: object
...
  resp.Resp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: 校灵通 API 文档
  title: EventGlide API
paths:
...
  /interaction/collect:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 互动
        in: body
        name: interaction
        required: true
        schema:
          $ref: '#/definitions/req.InteractionReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 收藏
      tags:
      - Interaction
  /interaction/discollect:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 互动
        in: body
        name: interaction
        required: true
        schema:
          $ref: '#/definitions/req.InteractionReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 取消收藏
      tags:
      - Interaction
  /interaction/dislike:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 互动
        in: body
        name: interaction
        required: true
        schema:
          $ref: '#/definitions/req.InteractionReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 取消点赞
      tags:
      - Interaction
  /interaction/like:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 互动
        in: body
        name: interaction
        required: true
        schema:
          $ref: '#/definitions/req.InteractionReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 点赞
      tags:
      - Interaction
  /post/all:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resp.ListPostsResp'
                  type: array
              type: object
      summary: 获取所有帖子
      tags:
      - Post
  /post/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/req.CreatePostReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 创建帖子
      tags:
      - Post
  /post/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 删除帖子
      tags:
      - Post
  /post/draft:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 草稿
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/req.CreatePostReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 创建草稿
      tags:
      - Post
  /post/find:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子名
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/req.FindPostReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resp.ListPostsResp'
                  type: array
              type: object
      summary: 通过帖子名查找帖子
      tags:
      - Post
  /post/load:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 草稿请求
        in: body
        name: draft
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/model.PostDraft'
              type: object
      summary: 加载草稿
      tags:
      - Post
  /post/own/{id}:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resp.ListPostsResp'
                  type: array
              type: object
      summary: 通过用户ID查找帖子
      tags:
      - Post
...
swagger: "2.0"